Split main component into individual components and CSS files.
