+++ src/styles/globals.css
@@ -2,6 +2,13 @@
 @tailwind components;
 @tailwind utilities;
 
+/* Main */
+@import 'main/header.css';
+@import 'main/login_mobie.css';
+@import 'main/mobile.css';
+
+@import 'main/header/filters.css';
+
 /* Mod Browser */
 @import 'modbrowser/cards.css';
 @import 'modbrowser/table.css';
@@ -24,6 +31,14 @@
 @import 'forms/content/mod.css';
 @import 'forms/content/category.css';
 @import 'forms/content/source.css';
+
+/* Main */
+main {
+    @apply flex;
+    @apply min-h-screen;
+    @apply flex-col;
+    @apply pb-20;
+}
 
 /* Background And Overlay */
 #bgol
