+++ src/styles/main/header/filters.css
@@ -1,0 +1,74 @@
+.header-filters {
+    @apply hidden;
+    @apply w-full;
+    @apply sm:flex;
+    @apply w-full;
+    @apply justify-center;
+    @apply items-center;
+    @apply gap-2;
+    @apply flex-wrap;
+}
+
+.header-filter-input {
+    @apply block;
+    @apply p-4;
+    @apply w-full;
+    @apply text-lg;
+    @apply text-gray-100;
+    @apply bg-gray-700;
+    @apply rounded-lg;
+    @apply border;
+    @apply border-gray-600;
+    @apply focus:ring-cyan-700;
+    @apply focus:border-cyan-700;
+}
+
+.header-filter-input-search {
+    @apply pl-10;
+    @apply !text-sm;
+}
+
+.header-filter-timeframe {
+    @apply relative;
+    @apply w-full;
+    @apply sm:w-32;
+}
+
+.header-filter-sort {
+    @apply relative;
+    @apply w-full;
+    @apply sm:w-3/12;
+}
+
+.header-filter-search {
+    @apply relative;
+    @apply w-full;
+    @apply sm:w-1/3;
+}
+
+.header-filter-button {
+    @apply relative;
+    @apply w-full;
+    @apply sm:w-16;
+    @apply flex;
+    @apply items-center;
+    @apply justify-center;
+    @apply bg-gray-700;
+    @apply rounded-lg;
+    @apply border;
+    @apply border-gray-600;
+    @apply hover:bg-cyan-700;
+}
+
+.header-filter-button > button > div, .header-filter-button > a > div {
+    @apply block;
+    @apply p-2;
+    @apply w-full;
+    @apply text-lg;
+    @apply text-gray-100;
+}
+
+.header-filter-button > button > div > svg, .header-filter-button > a > div > svg {
+    @apply w-10;
+    @apply h-10;
+}
