+++ src/styles/main/login_mobie.css
@@ -1,0 +1,5 @@
+#mobile-and-login {
+    @apply flex;
+    @apply flex-wrap;
+    @apply justify-between;
+}
